try again

+++ dict/migrations/0004_auto_20191126_1604.py
@@ -13,6 +13,6 @@ class Migration(migrations.Migration):
         migrations.AlterField(
             model_name='word',
             name='id',
-            field=models.AutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID'),
+            field=models.BigIntegerField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID'),
         ),
     ]
